10th commit

+++ scripts/ordersPage.js
@@ -7,8 +7,8 @@ const ordersTableTemplate = `
             <td>QUANTITY</td>
             <td>SHIP_DATE</td>
             <td>STATUS</td>
-            <td><button class="btn BUTTON_STYLE IS_DISABLED" onClick="ORDER_PROCESS(ORDER_ID)">BUTTON_COLOR</button>
-                <button class="btn btn-outline-danger" onClick="cancelOrder(ORDER_ID)">Cancel order</button>
+            <td><button class="btn CONFIRM_BUTTON_STYLE" onClick="ORDER_PROCESS(ORDER_ID)">CONFIRM_BUTTON_TEXT</button>
+                <button class="btn DELETE_BUTTON_STYLE" onClick="cancelOrder(ORDER_ID)">Cancel order</button>
             </td>
         </tr>
     `;
@@ -35,14 +35,14 @@ $(function () {
                 orderHTML = orderHTML.replace(/SHIP_DATE/g, order.shipDate)
                 orderHTML = orderHTML.replace(/STATUS/g, order.status)
                 if (order.status === "placed") {
-                    orderHTML = orderHTML.replace(/BUTTON_COLOR/g, "Wait for Shipping")
-                    orderHTML = orderHTML.replace(/BUTTON_STYLE/g, "btn-outline-primary")
-                    orderHTML = orderHTML.replace(/IS_DISABLED/g, "disabled")
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_TEXT/g, "Wait for Shipping")
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_STYLE/g, "btn-outline-primary disabled")
+                    orderHTML = orderHTML.replace(/DELETE_BUTTON_STYLE/g, "btn-danger")
                 }
                 else {
-                    orderHTML = orderHTML.replace(/BUTTON_COLOR/g, "Confirm receipt")
-                    orderHTML = orderHTML.replace(/BUTTON_STYLE/g, "btn-outline-success")
-                    orderHTML = orderHTML.replace(/IS_DISABLED/g, '')
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_TEXT/g, "Confirm receipt")
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_STYLE/g, "btn-success")
+                    orderHTML = orderHTML.replace(/DELETE_BUTTON_STYLE/g, "btn-outline-danger disabled")
                 }
 
                 count++;
@@ -55,12 +55,12 @@ $(function () {
 });
 
 function adminRenderTable(templete) {
+    let count = 1;
+    let ordersHTML = '';
     for (let i=0; i<localStorage.length; i++) {
         let key = localStorage.key(i);
-        let count = 1;
         if (key !== "username") {
             let ordersData = JSON.parse(localStorage.getItem(key));
-            let ordersHTML = '';
 
             ordersData.forEach((order) => {
                 let orderHTML = templete;
@@ -73,14 +73,14 @@ function adminRenderTable(templete) {
                 orderHTML = orderHTML.replace(/SHIP_DATE/g, order.shipDate)
                 orderHTML = orderHTML.replace(/STATUS/g, order.status)
                 if (order.status === "placed") {
-                    orderHTML = orderHTML.replace(/BUTTON_COLOR/g, "Confirm Shipping")
-                    orderHTML = orderHTML.replace(/BUTTON_STYLE/g, "btn-outline-success")
-                    orderHTML = orderHTML.replace(/IS_DISABLED/g, '')
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_TEXT/g, "Confirm Shipping")
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_STYLE/g, "btn-success")
+                    orderHTML = orderHTML.replace(/DELETE_BUTTON_STYLE/g, "btn-danger")
                 }
                 else {
-                    orderHTML = orderHTML.replace(/BUTTON_COLOR/g, "Product is shipping")
-                    orderHTML = orderHTML.replace(/BUTTON_STYLE/g, "btn-outline-primary")
-                    orderHTML = orderHTML.replace(/IS_DISABLED/g, 'disabled')
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_TEXT/g, "Product is shipping")
+                    orderHTML = orderHTML.replace(/CONFIRM_BUTTON_STYLE/g, "btn-outline-primary disabled")
+                    orderHTML = orderHTML.replace(/DELETE_BUTTON_STYLE/g, "btn-outline-danger disabled")
                 }
 
                 count++;
@@ -88,7 +88,6 @@ function adminRenderTable(templete) {
             })
 
             document.getElementById("orderTable").innerHTML = ordersHTML;
-
         }
     }
 }
@@ -102,47 +101,54 @@ function isExistOrdersForUser() {
 }
 
 function shipOrder(id) {
-    for (let i=0; i<localStorage.length; i++) {
-        let key = localStorage.key(i);
-        if (key !== "username") {
-            let ordersData = JSON.parse(localStorage.getItem(key));
-            for (let j=0; j<ordersData.length; j++) {
-                if (id === ordersData[j].id) {
-                    ordersData[j].status = "shipping";
-                    break;
+    if (confirm("Please confirm once again that you have packed and shipped it successfully.")) {
+        for (let i = 0; i < localStorage.length; i++) {
+            let key = localStorage.key(i);
+            if (key !== "username") {
+                let ordersData = JSON.parse(localStorage.getItem(key));
+                for (let j = 0; j < ordersData.length; j++) {
+                    if (id === ordersData[j].id) {
+                        ordersData[j].status = "shipping";
+                    }
                 }
+                localStorage.setItem(key, JSON.stringify(ordersData));
             }
-            localStorage.setItem(key, JSON.stringify(ordersData));
-            break;
         }
+        adminRenderTable(ordersTableTemplate);
+        bs4pop.notice('Order ' + id + ' shipped!',{type:'success'})
     }
-    adminRenderTable(ordersTableTemplate);
 }
 
 function completeOrder(id) {
-    if (isExistOrdersForUser()) {
-        let user = localStorage.getItem("username");
-        let ordersData = JSON.parse(localStorage.getItem(user));
-        ordersData = ordersData.filter((item)=>{
-            return item.id !== id;
-        });
-        localStorage.setItem(user, JSON.stringify(ordersData));
-        alert('Order ' + id + ' completed!')
-        document.getElementById(id).remove();
-        displayOrderQuantity();
+    if (confirm("Please confirm once again that you have received the package.")) {
+        if (isExistOrdersForUser()) {
+            let user = localStorage.getItem("username");
+            let ordersData = JSON.parse(localStorage.getItem(user));
+            ordersData = ordersData.filter((item) => {
+                return item.id !== id;
+            });
+            localStorage.setItem(user, JSON.stringify(ordersData));
+            document.getElementById(id).remove();
+            displayOrderQuantity();
+            bs4pop.notice('Order ' + id + ' completed!',{type:'success'})
+        }
     }
 }
 
 function cancelOrder(id) {
-    if (isExistOrdersForUser()) {
-        let user = localStorage.getItem("username");
-        let ordersData = JSON.parse(localStorage.getItem(user));
-        ordersData = ordersData.filter((item)=>{
-            return item.id !== id;
-        });
-        localStorage.setItem(user, JSON.stringify(ordersData));
-        alert('Order ' + id + ' deleted!')
+    if (confirm("Please confirm once again that you want to cancel the order.")) {
+        for (let i = 0; i < localStorage.length; i++) {
+            let key = localStorage.key(i);
+            if (key !== "username") {
+                let ordersData = JSON.parse(localStorage.getItem(key));
+                ordersData = ordersData.filter((item) => {
+                    return item.id !== id;
+                });
+                localStorage.setItem(key, JSON.stringify(ordersData));
+            }
+        }
         document.getElementById(id).remove();
         displayOrderQuantity();
+        bs4pop.notice('Order ' + id + ' cancelled!',{type:'success'})
     }
 }
